+++ packages/latex/src/index.ts
@@ -6,3 +6,4 @@ export * from "./compile-log/types";
 export * from "./compile-log/latex-log";
 export * from "./compile-log/bibtex-log";
 export * from "./compile-log/biber-log";
+export * from "./import-compat";